+++ src/services/trainer.js
@@ -20,7 +20,7 @@ const {
   levelConfig,
 } = require("../config/trainerConfig");
 const { logger } = require("../log");
-const { formatMoney, getFullUTCTimeInterval } = require("../utils/utils");
+const { formatMoney, setDefaultFields } = require("../utils/utils");
 const {
   backpackItems,
   backpackCategories,
@@ -102,68 +102,6 @@ const getTrainerFromId = async (userId) => {
 };
 
 /**
- *
- * @param {any} root
- * @param {TrainerFieldConfig} fieldConfig
- * @param {Date} lastCorrectedTime
- * @param {Date} newCorrectedTime
- * @returns {boolean} if field was modified
- */
-const setTrainerFields = (
-  root,
-  fieldConfig,
-  lastCorrectedTime,
-  newCorrectedTime
-) => {
-  let modified = false;
-  // check if all fields are present
-  for (const field in fieldConfig) {
-    const fieldData = fieldConfig[field];
-    if (root[field] === undefined) {
-      // eslint-disable-next-line no-param-reassign
-      root[field] = fieldData.default;
-      modified = true;
-    }
-
-    // if the field has its own config, attempt to set it
-    if (fieldData.type === "object" && fieldData.config) {
-      modified =
-        setTrainerFields(
-          root[field] ?? {},
-          fieldData.config,
-          lastCorrectedTime,
-          newCorrectedTime
-        ) || modified;
-    }
-  }
-
-  // attempt to reset time-interval fields
-  for (const field in fieldConfig) {
-    const { refreshInterval } = fieldConfig[field];
-    if (!refreshInterval) {
-      continue;
-    }
-
-    const currentTimeInterval = getFullUTCTimeInterval(
-      refreshInterval,
-      newCorrectedTime
-    );
-    const lastTimeInterval = getFullUTCTimeInterval(
-      refreshInterval,
-      lastCorrectedTime
-    );
-    if (currentTimeInterval > lastTimeInterval) {
-      // eslint-disable-next-line no-param-reassign
-      root[field] = fieldConfig[field].default;
-      modified = true;
-    }
-
-    return modified;
-  }
-};
-
-/**
- *
  * @param {{
  *  id: string,
  *  username?: string,
@@ -227,7 +165,7 @@ const getTrainer = async (discordUser, refresh = true) => {
   const lastCorrectedTime = new Date(trainer.lastCorrected);
   const newCorrectedTime = new Date();
   modified =
-    setTrainerFields(
+    setDefaultFields(
       trainer,
       trainerFields,
       lastCorrectedTime,